+++ dal/src/main/java/com/timevale/forward/dal/entity/ProductDemandGroupItemDO.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 /**
  * 产品-分组关系表实体类
  * 对应表：product_demand_group_item
@@ -19,7 +21,7 @@ public class ProductDemandGroupItemDO extends BaseDO {
     // 产品需求id
     private Long productDemandId;
     // 相对位置
-    private Double position;
+    private BigDecimal position;
     // 版本号
     private Long version;
     /**
